# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerCatalyst API",
    "description": "Multi-tenant AI-driven SaaS for ATS-optimized resume tailoring",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/profile/upload-resume": {
      "post": {
        "tags": [
          "Profile & Ingestion"
        ],
        "summary": "Upload and parse a resume document (PDF)",
        "description": "Upload a candidate resume PDF, extract text in-memory, and use Gemini to\nsynthesize a strictly typed ParsedProfile object.\nOptionally saves the candidate profile and ingests bullets into pgvector.",
        "operationId": "upload_and_parse_resume_api_profile_upload_resume_post",
        "parameters": [
          {
            "name": "save_to_db",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to immediately persist extracted profile and embed bullets in pgvector",
              "default": false,
              "title": "Save To Db"
            },
            "description": "Whether to immediately persist extracted profile and embed bullets in pgvector"
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_and_parse_resume_api_profile_upload_resume_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/save": {
      "post": {
        "tags": [
          "Profile & Ingestion"
        ],
        "summary": "Create or update candidate profile",
        "description": "Persist or update candidate profile info in the relational database.",
        "operationId": "save_profile_api_profile_save_post",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": [
          "Profile & Ingestion"
        ],
        "summary": "Retrieve candidate profile",
        "description": "Fetch candidate profile for the authenticated tenant and user.",
        "operationId": "get_profile_api_profile_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/bullets": {
      "post": {
        "tags": [
          "Profile & Ingestion"
        ],
        "summary": "Add and embed master bullets",
        "description": "Add new master bullets to the candidate vault and synchronize their\nembeddings into pgvector with tenant/user isolation.",
        "operationId": "add_bullets_api_profile_bullets_post",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MasterBulletCreate"
                },
                "title": "Bullets"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterBulletResponse"
                  },
                  "title": "Response Add Bullets Api Profile Bullets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Profile & Ingestion"
        ],
        "summary": "List all master bullets",
        "description": "Retrieve all master bullets associated with the current tenant and user.",
        "operationId": "list_bullets_api_profile_bullets_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_tenant",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "default_user",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterBulletResponse"
                  },
                  "title": "Response List Bullets Api Profile Bullets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_and_parse_resume_api_profile_upload_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_and_parse_resume_api_profile_upload_resume_post"
      },
      "CertificationItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "issuer": {
            "type": "string",
            "title": "Issuer"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "credential_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "issuer"
        ],
        "title": "CertificationItem",
        "description": "Professional certification record."
      },
      "EducationItem": {
        "properties": {
          "institution": {
            "type": "string",
            "title": "Institution"
          },
          "degree": {
            "type": "string",
            "title": "Degree"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "start_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year"
          },
          "gpa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpa"
          }
        },
        "type": "object",
        "required": [
          "institution",
          "degree"
        ],
        "title": "EducationItem",
        "description": "Academic degree / institution record."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasterBulletCreate": {
        "properties": {
          "bullet_text": {
            "type": "string",
            "title": "Bullet Text"
          },
          "skills_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Used",
            "default": []
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "Work Experience"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "impact_metrics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Metrics"
          },
          "experience_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Id"
          }
        },
        "type": "object",
        "required": [
          "bullet_text"
        ],
        "title": "MasterBulletCreate",
        "description": "Payload to create a new quantified achievement / master bullet point."
      },
      "MasterBulletResponse": {
        "properties": {
          "bullet_text": {
            "type": "string",
            "title": "Bullet Text"
          },
          "skills_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Used",
            "default": []
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "Work Experience"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "impact_metrics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Metrics"
          },
          "experience_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Id"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "bullet_text",
          "id",
          "tenant_id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "MasterBulletResponse",
        "description": "Master bullet response from database."
      },
...
      "ParsedExperienceItem": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "bullet_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullet Points",
            "default": []
          },
          "skills_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Used",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "company",
          "role"
        ],
        "title": "ParsedExperienceItem",
        "description": "Work experience item extracted from a candidate resume."
      },
      "ParsedProfile": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "linkedin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin"
          },
          "github": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github"
          },
          "portfolio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "type": "array",
            "title": "Education",
            "default": []
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ParsedExperienceItem"
            },
            "type": "array",
            "title": "Experiences",
            "default": []
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "type": "array",
            "title": "Projects",
            "default": []
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CertificationItem"
            },
            "type": "array",
            "title": "Certifications",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email"
        ],
        "title": "ParsedProfile",
        "description": "Output schema for resume document (PDF/DOCX) extraction via Gemini."
      },
      "ProfileCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "portfolio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "education": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Education",
            "default": []
          },
          "certifications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Certifications",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email"
        ],
        "title": "ProfileCreate",
        "description": "Payload to create or initialize a candidate profile."
      },
      "ProfileResponse": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "portfolio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "education": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Education",
            "default": []
          },
          "certifications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Certifications",
            "default": []
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "id",
          "tenant_id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileResponse",
        "description": "Complete candidate profile response from database."
      },
      "ProjectItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "technologies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technologies",
            "default": []
          },
          "bullet_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullet Points",
            "default": []
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProjectItem",
        "description": "Portfolio / personal project entry."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}